+++ class/actes/ActesTransactionsSQL.class.php
@@ -1,6 +1,7 @@
 <?php 
 class ActesTransactionsSQL extends SQL{
 
+	const MAX_ID = 2147483647; /* (signed) integer max size in PostgreSQL*/
 
 	public function getInfo($id){
 		$sql = "SELECT * FROM actes_transactions WHERE id=?";
@@ -357,5 +358,12 @@ class ActesTransactionsSQL extends SQL{
             $status_id
         );
     }
+
+    public function getAllForExport($authority_id,$min_transaction_id=0,$max_transaction_id=self::MAX_ID){
+		$sql = "SELECT actes_transactions.id, unique_id,file_path,file_size FROM actes_transactions " .
+			" JOIN actes_envelopes on actes_transactions.envelope_id = actes_envelopes.id " .
+			" WHERE actes_transactions.authority_id=? AND actes_transactions.id>=? AND actes_transactions.id<=?";
+		return $this->query($sql,$authority_id,$min_transaction_id,$max_transaction_id);
+	}
 
 }