+++ init/init.php
@@ -92,6 +92,7 @@ $objectInstancier->set("openstack_region",OPENSTACK_REGION);
 $objectInstancier->set("openstack_swift_container_prefix",OPENSTACK_SWIFT_CONTAINER_PREFIX);
 
 $objectInstancier->set("helios_files_upload_root",HELIOS_FILES_UPLOAD_ROOT);
+$objectInstancier->set("helios_responses_root",HELIOS_RESPONSES_ROOT);
 
 $objectInstancier->set("actes_files_upload_root",ACTES_FILES_UPLOAD_ROOT);
 $objectInstancier->set("actes_appli_trigramme",ACTES_APPLI_TRIGRAMME);