+++ test/PHPUnit/ActesUtilitiesTestTrait.php
@@ -14,7 +14,13 @@ trait ActesUtilitiesTestTrait
 	}
 
 
-	protected function createTransaction($status){
+	/**
+	 * @param $status
+	 * @param string $archive_path
+	 * @return array|bool|mixed
+	 * @throws Exception
+	 */
+	protected function createTransaction($status,$archive_path=""){
 		$sql="INSERT INTO actes_envelopes(user_id,siren,department) VALUES(1,'000000000','034') returning ID";
 		$envelope_id = $this->getSQLQuery()->queryOne($sql);
 
@@ -35,7 +41,18 @@ trait ActesUtilitiesTestTrait
 
 		$actesTransactionSQL->updateStatus($transaction_id,$status,"");
 
+		if ($archive_path){
+			$relative_path = basename($archive_path);
+			$destination = $this->getObjectInstancier()->get("actes_files_upload_root")."/".basename($archive_path);
+			copy($archive_path,$destination);
+			$sql = "UPDATE actes_envelopes SET file_path=?,file_size=? WHERE id=?";
+			$this->getSQLQuery()->query($sql,$relative_path,filesize($archive_path),$envelope_id);
+		}
 
+		$unique_id = $actesTransactionSQL->guessUniqueId($transaction_id);
+
+		$sql = "UPDATE actes_transactions SET unique_id=? WHERE id=?";
+		$this->getSQLQuery()->query($sql,$unique_id,$transaction_id);
 
 		return $transaction_id;
 	}